Add Variant#display_images
When displaying Images for a Variant, we want to include Images
associated with the Product's master Variant as well. These Images are
considered applicable to all Variants of the Product. This change allows
us to build that list of Images via arel without disrupting the use of
the #images association.

One other change required here is that in some cases we need to allow
Images to be created without having a Variant associated with them. We
add a validation on VariantImage to prevent invalid join records from
being created.