Remove most order() calls
In Rails 5.1, SQL queries with DISTINCT, COUNT(), and ORDER_BY are not
parsed correctly leading to an error from Postgres. A similar problem
was reported and fixed on

https://github.com/rails/rails/pull/29848

but I don't think it is the same problem as this one.

I prefer to avoid putting ORDER BY clauses into has_many associtainos,
default scopes, & etc. because they inevitably get in the way. By
removing them here, we will probably have to add  them back in a
presenter later. For now it's enough to get them out of the way.